+++ src/TextureManager.cpp
@@ -12,6 +12,7 @@
 #include <stdarg.h>
 
 #include "global.h"
+#include "OpenRaider.h"
 #include "utils/pcx.h"
 #include "utils/pixel.h"
 #include "utils/strings.h"
@@ -19,45 +20,37 @@
 #include "TextureManager.h"
 
 TextureManager::TextureManager() {
-    mTextureIds = NULL;
     mFlags = 0;
-    mTextureCount = 0;
-    mTextureLimit = 0;
 }
 
 TextureManager::~TextureManager() {
     reset();
 }
 
-unsigned char *TextureManager::generateColorTexture(const unsigned char rgba[4],
-        unsigned int width, unsigned int height) {
-    assert(rgba != NULL);
-    assert(width > 0);
-    assert(height > 0);
-
-    unsigned char *image = new unsigned char[height * width * 4];
-
-    for (unsigned int i = 0; i < (width * height); i++) {
-        image[i * 4] = rgba[0];
-        image[(i * 4) + 1] = rgba[1];
-        image[(i * 4) + 2] = rgba[2];
-        image[(i * 4) + 3] = rgba[3];
+void TextureManager::reset() {
+    while (mTextureIds.size() > 0) {
+        unsigned int id = mTextureIds.at(mTextureIds.size() - 1);
+        glDeleteTextures(1, &id);
+        mTextureIds.pop_back();
     }
-
-    return image;
 }
 
-int TextureManager::loadColorTexture(const unsigned char rgba[4],
-        unsigned int width, unsigned int height) {
-    assert(rgba != NULL);
-    assert(width > 0);
-    assert(height > 0);
-
-    unsigned char *image = generateColorTexture(rgba, width, height);
-    int id = loadBuffer(image, width, height, RGBA, 32);
+int TextureManager::initialize() {
+    unsigned char *image = generateColorTexture(WHITE, 32, 32, 32);
+    loadBufferSlot(image, 32, 32, RGBA, 32, mTextureIds.size());
     delete [] image;
 
-    return id;
+    //! \fixme Temporary
+    char *filename = bufferString("%s/tr2/TITLE.PCX", getOpenRaider().mPakDir);
+    if (loadPCX(filename) < 0) {
+        delete [] filename;
+        //! \fixme Error Checking. Return negative error code, check in calling place too
+        filename = bufferString("%s/%s", getOpenRaider().mDataDir, "splash.tga");
+        loadTGA(filename);
+    }
+    delete [] filename;
+
+    return (mTextureIds.size() == 0) ? -1 : 0;
 }
 
 void TextureManager::setFlag(TextureFlag flag) {
@@ -68,15 +61,8 @@ void TextureManager::clearFlag(TextureFlag flag) {
     mFlags &= ~flag;
 }
 
-void TextureManager::reset() {
-    if (mTextureIds) {
-        glDeleteTextures(mTextureLimit, mTextureIds);
-        delete [] mTextureIds;
-    }
-
-    mTextureIds = NULL;
-    mTextureCount = 0;
-    mTextureLimit = 0;
+int TextureManager::getTextureCount() {
+    return mTextureIds.size();
 }
 
 void TextureManager::disableMultiTexture() {
@@ -94,128 +80,94 @@ void TextureManager::useMultiTexture(float aU, float aV, float bU, float bV) {
     glMultiTexCoord2fARB(GL_TEXTURE1_ARB, bU, bV);
 }
 
-void TextureManager::useMultiTexture(float u, float v) {
-    useMultiTexture(u, v, u, v);
-}
-
 void TextureManager::bindMultiTexture(int texture0, int texture1) {
-    assert(mTextureIds != NULL);
     assert(texture0 >= 0);
     assert(texture1 >= 0);
-    assert((unsigned int)texture0 <= mTextureCount);
-    assert((unsigned int)texture1 <= mTextureCount);
+    assert((unsigned int)texture0 < mTextureIds.size());
+    assert((unsigned int)texture1 < mTextureIds.size());
 
     mFlags |= fUseMultiTexture;
 
     glActiveTextureARB(GL_TEXTURE0_ARB);
     glEnable(GL_TEXTURE_2D);
-    glBindTexture(GL_TEXTURE_2D, mTextureIds[texture0]);
+    glBindTexture(GL_TEXTURE_2D, mTextureIds.at(texture0));
 
     glActiveTextureARB(GL_TEXTURE1_ARB);
     glEnable(GL_TEXTURE_2D);
-    glBindTexture(GL_TEXTURE_2D, mTextureIds[texture1]);
-}
-
-void TextureManager::setMaxTextureCount(unsigned int n) {
-    mTextureLimit = n;
-    mTextureIds = new unsigned int[n];
-    glGenTextures(n, mTextureIds);
-}
-
-int TextureManager::getTextureCount() {
-    return mTextureCount - 1;
-}
-
-int TextureManager::loadBuffer(unsigned char *image,
-        unsigned int width, unsigned int height,
-        ColorMode mode, unsigned int bpp) {
-    int id;
-
-    assert(image != NULL);
-    assert(width > 0);
-    assert(height > 0);
-    assert((bpp == 8) || (bpp == 24) || (bpp == 32));
-
-    id = loadBufferSlot(image, width, height, mode, bpp, mTextureCount++);
-
-    if (id < 0)
-        return id;
-
-    return id;
+    glBindTexture(GL_TEXTURE_2D, mTextureIds.at(texture1));
 }
 
 int TextureManager::loadBufferSlot(unsigned char *image,
         unsigned int width, unsigned int height,
         ColorMode mode, unsigned int bpp,
         unsigned int slot) {
-    unsigned char bytes;
-    unsigned int glcMode;
-
-    assert(mTextureIds != NULL);
-    assert(slot < mTextureLimit);
     assert(image != NULL);
     assert(width > 0);
     assert(height > 0);
-    assert((bpp == 8) || (bpp == 24) || (bpp == 32));
+    assert((mode == GREYSCALE) || (mode == RGB)
+            || (mode == BGR) || (mode == ARGB)
+            || (mode == RGBA) || (mode ==  BGRA));
 
     switch (mode) {
         case GREYSCALE:
-            if (bpp != 8) {
-                printf("TextureManager::Load ERROR Unsupported GREYSCALE, %i bpp\n", bpp);
-                return -1;
-            }
+            assert(bpp == 8);
+            break;
+
+        case RGB:
+        case BGR:
+            assert(bpp == 24);
+            break;
+
+        case ARGB:
+        case RGBA:
+        case BGRA:
+            assert(bpp == 32);
+            break;
+    }
+
+    while (mTextureIds.size() <= slot) {
+        unsigned int id;
+        glGenTextures(1, &id);
+        mTextureIds.push_back(id);
+    }
+
+    unsigned char bytes;
+    unsigned int glcMode;
+
+    switch (mode) {
+        case GREYSCALE:
             bytes = 1;
             glcMode = GL_LUMINANCE;
             break;
 
         case RGB:
-            if (bpp != 24) {
-                printf("TextureManager::Load ERROR Unsupported RGB, %i bpp\n", bpp);
-                return -1;
-            }
             bytes = 3;
             glcMode = GL_RGB;
             break;
 
         case ARGB:
-            if (bpp == 32) {
-                argb2rgba32(image, width, height);
-            } else {
-                printf("TextureManager::Load ERROR Unsupported ARGB, %i bpp\n", bpp);
-                return -1;
-            }
+            argb2rgba32(image, width, height);
             bytes = 4;
             glcMode = GL_RGBA;
             break;
 
         case RGBA:
-            if (bpp != 32) {
-                printf("TextureManager::Load ERROR Unsupported RGBA, %i bpp\n", bpp);
-                return -1;
-            }
             bytes = 4;
             glcMode = GL_RGBA;
             break;
 
         case BGR:
-            if (bpp != 24) {
-                printf("TextureManager::Load ERROR Unsupported BGR, %i bpp\n", bpp);
-                return -1;
-            }
             bytes = 3;
             glcMode = GL_BGR_EXT;
             break;
 
         case BGRA:
-            if (bpp != 32) {
-                printf("TextureManager::Load ERROR Unsupported BGRA, %i bpp\n", bpp);
-                return -1;
-            }
             bytes = 4;
             glcMode = GL_BGRA_EXT;
             break;
     }
 
+    glColor3ubv(WHITE);
     glClearColor(0.0, 0.0, 0.0, 0.0);
     glEnable(GL_DEPTH_TEST);
     glShadeModel(GL_SMOOTH);
@@ -241,13 +193,12 @@ int TextureManager::loadBufferSlot(unsigned char *image,
 }
 
 void TextureManager::bindTextureId(unsigned int n) {
-    assert(mTextureIds != NULL);
-    assert(n <= mTextureCount);
+    assert(n < mTextureIds.size());
 
     glEnable(GL_TEXTURE_2D);
     //glTexEnvf(GL_TEXTURE_ENV, GL_TEXTURE_ENV_MODE, GL_MODULATE);
 
-    glBindTexture(GL_TEXTURE_2D, mTextureIds[n]);
+    glBindTexture(GL_TEXTURE_2D, mTextureIds.at(n));
 }
 
 int TextureManager::loadPCX(const char *filename) {
@@ -266,7 +217,7 @@ int TextureManager::loadPCX(const char *filename) {
             delete [] image;
             image = image2;
         }
-        id = loadBuffer(image, w, h, c, bpp);
+        id = loadBufferSlot(image, w, h, c, bpp, mTextureIds.size());
         delete [] image;
     }
 
@@ -277,8 +228,7 @@ int TextureManager::loadTGA(const char *filename) {
     assert(filename != NULL);
     assert(filename[0] != '\0');
 
-    unsigned char *image = NULL;
-    unsigned char *image2 = NULL;
+    unsigned char *image;
     unsigned int w, h;
     char type;
     int id = -1;
@@ -286,24 +236,18 @@ int TextureManager::loadTGA(const char *filename) {
     if (!tgaCheck(filename)) {
         tgaLoad(filename, &image, &w, &h, &type);
 
-        image2 = scaleBuffer(image, &w, &h, (type == 2) ? 32 : 24);
-
+        unsigned char *image2 = scaleBuffer(image, &w, &h, (type == 2) ? 32 : 24);
         if (image2) {
             delete [] image;
             image = image2;
         }
-
         if (image) {
-            id = loadBuffer(image, w, h,
+            id = loadBufferSlot(image, w, h,
                     (type == 2) ? RGBA : RGB,
-                    (type == 2) ? 32 : 24);
-
+                    (type == 2) ? 32 : 24,
+                    mTextureIds.size());
             delete [] image;
         }
-    }
-
-    if (id == -1) {
-        printf("TextureManager::loadTGA> ERROR: Failed to load '%s'\n", filename);
     }
 
     return id;